requirements.txt
# Core Framework
Flask==3.0.2
Werkzeug==3.0.1

# Production Server
gunicorn==21.2.0

# Database
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
SQLAlchemy==2.0.27
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
Flask-Login==0.6.3
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
bcrypt==4.1.2
PyJWT==2.8.0
email_validator==2.1.1

# Forms & Validation
Flask-WTF==1.2.1
WTForms==3.1.2

# API Documentation
flasgger==0.9.7.1

# Task Scheduling
Flask-APScheduler==1.13.1
APScheduler==3.10.4

# Security
Flask-Talisman==0.8.1

# Environment Management
python-dotenv==1.0.1

# Utilities
blinker==1.7.0
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
python-dateutil==2.8.2
pytz==2024.1
six==1.16.0